Reverted a change in the previous commit.
There is no mertit for making _RcpMaxBlurRadius half-float.

+++ Assets/Kino/Motion/Shader/Common.cginc
@@ -58,7 +58,7 @@ float2 _TileMaxOffs;
 
 // Maximum blur radius (in pixels)
 half _MaxBlurRadius;
-half _RcpMaxBlurRadius;
+float _RcpMaxBlurRadius;
 
 // Filter parameters/coefficients
 half _LoopCount;
